Separate the dialog from the restaurnat
